refactor: extract shared ShopifyProductCard for home + future store (#2)
Replace FeaturedProducts' inline product card with a single reusable
ShopifyProductCard built on ShopifyProduct. It's a superset of the legacy
Product-based ProductCard — the detail link, star rating and explicit
badge render only when the matching optional fields are present — so the
store pages can adopt this same component when they move onto
ShopifyProduct, with no behavioural drift between the two.

Home add-to-cart now matches the store card: real useCart().addItem with
the global CartNotification, dropping the local "Added!" placeholder.
